Only apply Windows-specific PoSh aliases if we're running on Windows

+++ posh/Aliases.ps1
@@ -5,43 +5,46 @@ Set-Alias -Name gita    -Value Invoke-GitChildDir
 Set-Alias -Name up      -Value Update-Profile
 Set-Alias -Name which   -Value Get-Command
 
-# Remove the curl alias if the real deal is present
-if (Get-Command -Name curl.exe -ErrorAction Ignore) {
-    Remove-Item -Path Alias:\curl -ErrorAction Ignore
-}
-
-# Remove the sc alias in favour of the sc.exe utility
-if (Get-Command -Name sc.exe -ErrorAction Ignore) {
-    Remove-Item -Path Alias:\sc -ErrorAction Ignore
-}
-
-# Dependency Walker (x86)
-$DepWalker32Path = Join-Path -Path ${env:ProgramFiles(x86)} -ChildPath 'Nexiom\Software\Independent\Dependency Walker\depends.exe'
-if (Test-Path -Path $DepWalker32Path -PathType Leaf) {
-    Set-Alias -Name depends32 -Value $DepWalker32Path
-} else {
-    Write-Verbose -Message '[dotfiles] Unable to locate Dependency Walker (x86) for alias.'
-}
-Remove-Variable -Name DepWalker32Path
-
-# Dependency Walker (x64)
-$DepWalker64Path = Join-Path -Path $env:ProgramW6432 -ChildPath 'Nexiom\Software\Independent\Dependency Walker\depends.exe'
-if (Test-Path -Path $DepWalker64Path -PathType Leaf) {
-    Set-Alias -Name depends64 -Value $DepWalker64Path
-} else {
-    Write-Verbose -Message '[dotfiles] Unable to locate Dependency Walker (x64) for alias.'
-}
-Remove-Variable -Name DepWalker64Path
-
-# Sublime Text 2
-$Sublime2RegPath = 'HKLM:Software\Microsoft\Windows\CurrentVersion\Uninstall\Sublime Text 2_is1'
-if (!(Get-Command -Name subl.exe -ErrorAction Ignore)) {
-    if (Test-Path -Path $Sublime2RegPath -PathType Container) {
-        Set-Alias -Name subl -Value (Join-Path -Path (Get-ItemProperty -Path $Sublime2RegPath).InstallLocation -ChildPath 'sublime_text.exe')
-    } else {
-        Write-Verbose -Message '[dotfiles] Unable to locate Sublime Text 2 for alias.'
+# Windows only
+if ($PSVersionTable.PSEdition -eq 'Desktop' -or $PSVersionTable.Platform -eq 'Win32NT') {
+    # Remove the curl alias if the real deal is present
+    if (Get-Command -Name curl.exe -ErrorAction Ignore) {
+        Remove-Item -Path Alias:\curl -ErrorAction Ignore
     }
-} else {
-    Write-Verbose -Message '[dotfiles] Skipping Sublime Text 2 alias as found subl.exe in PATH.'
+
+    # Remove the sc alias in favour of the sc.exe utility
+    if (Get-Command -Name sc.exe -ErrorAction Ignore) {
+        Remove-Item -Path Alias:\sc -ErrorAction Ignore
+    }
+
+    # Dependency Walker (x86)
+    $DepWalker32Path = Join-Path -Path ${env:ProgramFiles(x86)} -ChildPath 'Nexiom\Software\Independent\Dependency Walker\depends.exe'
+    if (Test-Path -Path $DepWalker32Path -PathType Leaf) {
+        Set-Alias -Name depends32 -Value $DepWalker32Path
+    } else {
+        Write-Verbose -Message '[dotfiles] Unable to locate Dependency Walker (x86) for alias.'
+    }
+    Remove-Variable -Name DepWalker32Path
+
+    # Dependency Walker (x64)
+    $DepWalker64Path = Join-Path -Path $env:ProgramW6432 -ChildPath 'Nexiom\Software\Independent\Dependency Walker\depends.exe'
+    if (Test-Path -Path $DepWalker64Path -PathType Leaf) {
+        Set-Alias -Name depends64 -Value $DepWalker64Path
+    } else {
+        Write-Verbose -Message '[dotfiles] Unable to locate Dependency Walker (x64) for alias.'
+    }
+    Remove-Variable -Name DepWalker64Path
+
+    # Sublime Text 2
+    $Sublime2RegPath = 'HKLM:Software\Microsoft\Windows\CurrentVersion\Uninstall\Sublime Text 2_is1'
+    if (!(Get-Command -Name subl.exe -ErrorAction Ignore)) {
+        if (Test-Path -Path $Sublime2RegPath -PathType Container) {
+            Set-Alias -Name subl -Value (Join-Path -Path (Get-ItemProperty -Path $Sublime2RegPath).InstallLocation -ChildPath 'sublime_text.exe')
+        } else {
+            Write-Verbose -Message '[dotfiles] Unable to locate Sublime Text 2 for alias.'
+        }
+    } else {
+        Write-Verbose -Message '[dotfiles] Skipping Sublime Text 2 alias as found subl.exe in PATH.'
+    }
+    Remove-Variable -Name Sublime2RegPath
 }
-Remove-Variable -Name Sublime2RegPath
